Query github for the latest version of docker-compose
Query github for the latest version of docker-compose.  This
functionality is achieved by calling the github api which returns a json
string of all available versions.  This json string is parsed with a
docker image (pinterb/jq).  If the image is not currently installed it
will automatically be downloaded.  Once the json string is parsed and
only the name of the release is parsed out then the sort --version-sort
command is run.  This will sort version numbers correctly.  Only the
last version is captured with "tail -1".  That version number is used
to download the latest version of docker-compose.