Cleanup docker image pinterb/jq
Once the json string for the docker-compose version is done being parsed
there is no need to keep the pinterb/jq image.  It is deleted at the end
of the upgradeDockerCompose function.  The image is small enough that it
can be downloaded each time the script is executed

Cleaned up the TODO messages about needing to fix the hard-coded
docker-compose version.