Add sudo ls as first step
Added the command "sudo ls" as the very first step of the script.  This
ensures the sudo password is set for the rest of the script execution.
Later on in the script there is apt-get commands that need access to the
sudo password.

If/when this script is converted to an ansible playbook it is much
simplier to supply the sudo password by using the ansible-playbook flag
--become-password